"Unpack" top-level elements with anonymous types
Consider the following snippet

	<s:element name="RecibeCFD">
	  <s:complexType>
	    <s:sequence>
	      <s:element minOccurs="0" maxOccurs="1" name="XMLCFD" type="s:string" />
	    </s:sequence>
	  </s:complexType>
	</s:element>

The `xsd` package would create an anonymous type name for that inner `complexType`,
such as `_anon1`. However, the accompanying wsdl document may have something
like

	<wsdl:part name="foo" type="RecibeCFD" />

The WSDL/XSD specs are so complicated that I can only assume that this is legal.
Therefore, this change duplicates the containing element name on such types.