Support nillable elements
The `nillable="true"` attribute in xsd elements maps quite nicely to
the `encoding/xml` package's `,omitempty` option in struct tags. With
this change, if a nillable value is the zero value for its type (e.g. 0 for int,
false for bool, "" for string, etc), it will not be present in the xml message.

This could come back to bite us, since Go's concept of zero values may
not always line up with that of an xml schema documents', but I think this
change is worth it.