Adjust the selected namespaces based on wsdl operations
The registerXSDTypes function would whitelist all types mentioned
in a WSDL operation's parts. However, it would not extend the namespaces
if the WSDL operation parts used types outside of the target name space.
So if a WSDL operation used `xsd:dateTime` directly, the `xsdgen` package would
not generate the helper types and methods for it! This commit corrects the
erroneous behavior.