Introduced wrapper structs when argument list grows too long
Once a function grows past a certain number of arguments, say 4,
it can be difficult to remember which order they come in, especially
if the function has a poor name. With this change, once a function
grows past a user-configurable threshold of input arguments and/or
return values, an additional type is introduced to hold all of the values.

As of now, the wrapper struct is passed by value. We may be better
off passing a pointer, especially if the struct is very large.