Handle mixed content types better.
Many complex types wrap simple types and add attributes to them.
Code generated for *any* complex types that are derived from such types
need to provide a ",chardata" field for the mixed content. Previously
the xsdgen package only generated such a field when a type was being
extended, but a restriction is just as valid.