Fix infinite loop in xmltree.diffScope
The diffScope function was used to get the xmlns declarations of an
element that differed from its parent's. It made the assumption that the
parent scope was always a prefix of the child scope. This is generally
true for any tree returned by xmltree.Parse.

However, once we begin modifying an xml tree post-Parse, that assumption
goes out the window. This can happen when the `xsd.derefGroupAttr`
function dereferences a link to an element in a separate schema.