Handle <choice> model groups with recurring elements
This is in response to issue #37. To handle `complexContent` like the following:

	<choice>
	  <sequence>
	    <element name="PGPKeyID" type="base64Binary"/>
	    <element name="PGPKeyPacket" type="base64Binary" minOccurs="0"/>
	    <any namespace="##other" processContents="lax" minOccurs="0"
	     maxOccurs="unbounded"/>
	  </sequence>
	  <sequence>
	    <element name="PGPKeyPacket" type="base64Binary"/>
	    <any namespace="##other" processContents="lax" minOccurs="0"
	     maxOccurs="unbounded"/>
	  </sequence>
	</choice>

Rather than try to capture some concept of sum types, which can be
cumbersome in Go code, we instead merge like-named elements
together, using some heuristics based on the marshal/unmarshal
behavior of the `encoding/xml` package.

Also fixed a bug in MarshalXML/UnmarshalXML methods for complexTypes
with duplicate element/attribute names.