Fix bug when modifying the last type declaration in a schema.
In the original code, there were two calls to

	root.Children = append(root.Children, t)

Because the `Children` field is a `[]xmltree.Element`, rather than an `[]*xmltree.Element`,
if the call to `append()` causes a reallocation, of the underlying array, then the `el` pointer
used within the same loop will no longer point to the current value, and updates to it will
be lost, as the added test demonstrated.

A better change would be to factor out the rearrangement of XMl such as

	<element name="foo">
	  <complexType name="foo">
	    ...
	  </complexType>
	</element>

to

	<element name="foo" type="foo"/>
	<complexType name="foo">
	  ...
	</complexType>

in a separate pass, rather than doing it in two places (the
`copyEltNamesToAnonymousTypes` and `nameAnonymousTypes` functions).